A coding of unfolding.
Expansion of rules where there are primitive operations operating on several variables that could
be transformed in primitives operating on a single variable when the other can be approximated by
the value they can take. This kind of expansion increases the number of rules but make them
manageable by Z3 with DifferenceOfCube a representation that scales well for network encoding.

It is an implementation of Pierre-Leo Begay's algorithm.

Co-authored-by: pierre.cregut@orange.com
Co-authored-by: pierreleo.begay@orange.com