Translate bool to bitvect of size 1
Translate the datalog bool type to a bitvector of size 1 instead of
the builtin bool type of z3 because the use of bool with bitvectors
result in core dumps when the doc mode is used.

Note: bool could not be used to interpret a variable as an atom, so the
change has no impact on octant datalog.