correct negation of atoms in the edb
Atoms in the edb are expressed with labelled args. The number of column
used will be expanded so that we retrieve from the backing database only
the columns we really need.
When done on a negated atom, we must first project the added columns
away and keep only the columns on which the negation lies.

This is done early in the compilation path (before typing). Type errors
may be harder to understand.

Closes #4